refactor of the class model to able pipeline
