+++ src/hipp/kh9pc/restitution/__init__.py
@@ -1,9 +1,9 @@
-from .base import RectificationStrategy
+from .base import RectificationStrategy, QCMixin
 from .flat_rectification_strategy import FlatRectificationStrategy
 from .poly_rectification_strategy import PolyRectificationStrategy
 from .collimation_rectification_strategy import CollimationRectificationStrategy
-
 from .vertical_edges_estimator import VerticalEdgesEstimator
+from .transformer import ImageTransformer, ImageTransformerTps, ImageTransformerAffine
 
 
 __all__ = [
@@ -12,4 +12,8 @@ __all__ = [
     "PolyRectificationStrategy",
     "CollimationRectificationStrategy",
     "VerticalEdgesEstimator",
+    "QCMixin",
+    "ImageTransformer",
+    "ImageTransformerTps",
+    "ImageTransformerAffine",
 ]
